[cif] refactor cif model, use ASRModel as parent (#1791)
* [cif] refactor cif model, use ASRModel as parent

* fix lint
